The main aim of this project was to build a Digital Communication system using RS-485 bus using Carrier Sense Multiple Access (CSMA) Technology.
The main aim of this project was to build a Digital Communication system using RS-485 bus using Carrier Sense Multiple Access (CSMA) Technology.
The microcontroller used was TM4C123GH6PM and the programming language was C, the hardware was built such that on a 2 wire RS-485 bus each Microcontroller acts as a node with CSMA Technology enabled and, these microcontroller nodes can communicate with each other effectively without interference and data loss.
The programming code was upto 2000 effective lines.